backend/requirements.txt
fastapi
uvicorn
sqlalchemy
psycopg2-binary
pydantic
email-validator
python-multipart
transformers
torch
sentence-transformers
openai
openai-whisper
nltk
numpy
scikit-learn
SpeechRecognition
pydub
librosa
python-jose[cryptography]
passlib[bcrypt]
PyPDF2
python-docx